+++ tofuref/data/registry.py
@@ -3,6 +3,7 @@ from typing import Dict, TYPE_CHECKING, Optional
 
 if TYPE_CHECKING:
     from tofuref.data.providers import Provider
+    from tofuref.data.resources import Resource
 
 
 @dataclass
@@ -12,6 +13,7 @@ class Registry:
     fullscreen_mode: bool = False
     providers: Dict[str, "Provider"] = field(default_factory=dict)
     active_provider: Optional["Provider"] = None
+    active_resource: Optional["Resource"] = None
 
 
 registry = Registry()